extract sns transfer to component
